+++ components/qrcode/src/call.ts
@@ -1,36 +1,17 @@
-import type { AnyFC, MaybeArray } from './types'
+import type { MaybeArray } from './types'
 
-function call(funcs: MaybeArray<() => void>): void
+/**
+ * Call one or multiple functions with the same arguments
+ */
+export function call<T extends (...args: any[]) => any>(
+  funcs: MaybeArray<T>,
+  ...args: Parameters<T>
+): void {
+  if (!funcs) return
 
-function call<A1>(funcs: MaybeArray<(a1: A1) => void>, a1: A1): void
-
-function call<A1, A2>(
-  funcs: MaybeArray<(a1: A1, a2: A2) => void>,
-  a1: A1,
-  a2: A2,
-): void
-
-function call<A1, A2, A3>(
-  funcs: MaybeArray<(a1: A1, a2: A2, a3: A3) => void>,
-  a1: A1,
-  a2: A2,
-  a3: A3,
-): void
-
-function call<A1, A2, A3, A4>(
-  funcs: MaybeArray<(a1: A1, a2: A2, a3: A3, a4: A4) => void>,
-  a1: A1,
-  a2: A2,
-  a3: A3,
-  a4: A4,
-): void
-
-function call<A extends any[]>(funcs: AnyFC[] | AnyFC, ...args: A) {
   if (Array.isArray(funcs)) {
-    funcs.forEach((func) => (call as any)(func, ...args))
+    funcs.forEach((func) => func?.(...args))
   } else {
-    return funcs(...args)
+    funcs(...args)
   }
 }
-
-export { call }